deploy fix #3

+++ src/components/Product/index.js
@@ -68,7 +68,7 @@ class Product extends Component{
 
         this.props.compare();
 
-        if(this.props.comparedItems.some(item => item.title.includes(product.title))){
+        if(this.props.comparedItems && this.props.comparedItems.some(item => item.title.includes(product.title))){
             console.log("remove from compare called");
             console.log("remove called on: ", product);
             store.dispatch(removeFromCompare(product));
@@ -97,39 +97,73 @@ class Product extends Component{
         store.dispatch(removeFromWatch(item));
     }
 
-    render(){
+    renderCard(){
+        if(this.props.comparedItems){
+            return(
+                <div className={" " + (this.props.comparedItems.some(item => item.title.includes(this.props.item.title)) ? "compare" : "")} >
+                    <div className="card productCard" >
+                            <div className="card-image itemImage" style={{display: "flex", justifyContent: "center"}}>
+                                    <img src={this.props.item.image} alt={this.props.item.title} style={{height:"250px", width: "60%", marginTop: "25px"}}/>
+                                    <div className="btnTool">
+                                        <span className={!this.props.item.watch? "btnTooltipText" : ""}>{!this.props.item.watch? "Add to Watchlist" : ""}</span>
+                                        <a className="btn-floating halfway-fab floatingWatchBtn indigo darken-4"><i className="material-icons" style={{color: (this.props.item.watch? "red" : "white")}} onClick={(this.props.item.watch) ? () => console.log("If you would like to remove this item from your watchlist, please remove it through the watchlist tab") : () => this.handleWatch(this.props.watch, this.props.item)}>{this.props.item.watch? "favorite" : "remove_red_eye"} </i></a>
+                                    </div>             
+                                <div className={(this.props.comparedItems.some(item => item.title.includes(this.props.item.title)) ? "card-overlay2" : "card-overlay")}></div>
+                                <div className="detailsBtn" onClick={() => this.handleCompare(this.props.item)} style={{color: "black", display: "flex", justifyContent: "center", alignItems:"center"}}>{this.props.comparedItems.some(item => item.title.includes(this.props.item.title)) ? "Remove Compare" : "Compare"}</div>
+                            </div>
+                            {/* this.props.item.featured && */}
+                            {/* <div className="featuredProduct">
+                                Featured
+                            </div> */}
+                            <div className="card-content" style={{display: "flex", alignItems: "center", flexDirection: "column"}}>
+                                <div className="card-title" style={{ lineHeight:"1.3em" ,fontSize:"18px", overflow:"hidden", position: "relative", bottom: "30%", textAlign: "center"}}>{this.props.item.title}</div>
+                                <div className="itemPrice">{`$${this.props.item.price}`}</div>
+                                <div>
+                                    <a href={`${this.props.item.link}`}  target="_blank" className="productSourceLogo"><img className={"img-fluid" + (this.props.item.logo == "https://firebasestorage.googleapis.com/v0/b/poofapibackend.appspot.com/o/icons%2Fbestbuy.png?alt=media&token=da783e57-802f-4a40-b609-3af1fd78098a" ? " productBestBuyLogo" : " productSourceLogoPic")} src={this.props.item.logo} alt={this.props.item.title}/></a>
+                                </div>
+                            </div>
+                    </div>
+                </div>
+            )
+        }
+        else{
+            return(
+                <div className={" "} >
+                    <div className="card productCard" >
+                            <div className="card-image itemImage" style={{display: "flex", justifyContent: "center"}}>
+                                    <img src={this.props.item.image} alt={this.props.item.title} style={{height:"250px", width: "60%", marginTop: "25px"}}/>
+                                    <div className="btnTool">
+                                        <span className={!this.props.item.watch? "btnTooltipText" : ""}>{!this.props.item.watch? "Add to Watchlist" : ""}</span>
+                                        <a className="btn-floating halfway-fab floatingWatchBtn indigo darken-4"><i className="material-icons" style={{color: (this.props.item.watch? "red" : "white")}} onClick={(this.props.item.watch) ? () => console.log("If you would like to remove this item from your watchlist, please remove it through the watchlist tab") : () => this.handleWatch(this.props.watch, this.props.item)}>{this.props.item.watch? "favorite" : "remove_red_eye"} </i></a>
+                                    </div>             
+                                <div className="card-overlay"></div>
+                                <div className="detailsBtn" onClick={() => this.handleCompare(this.props.item)} style={{color: "black", display: "flex", justifyContent: "center", alignItems:"center"}}>{"Compare"}</div>
+                            </div>
+                            {/* this.props.item.featured && */}
+                            {/* <div className="featuredProduct">
+                                Featured
+                            </div> */}
+                            <div className="card-content" style={{display: "flex", alignItems: "center", flexDirection: "column"}}>
+                                <div className="card-title" style={{ lineHeight:"1.3em" ,fontSize:"18px", overflow:"hidden", position: "relative", bottom: "30%", textAlign: "center"}}>{this.props.item.title}</div>
+                                <div className="itemPrice">{`$${this.props.item.price}`}</div>
+                                <div>
+                                    <a href={`${this.props.item.link}`}  target="_blank" className="productSourceLogo"><img className={"img-fluid" + (this.props.item.logo == "https://firebasestorage.googleapis.com/v0/b/poofapibackend.appspot.com/o/icons%2Fbestbuy.png?alt=media&token=da783e57-802f-4a40-b609-3af1fd78098a" ? " productBestBuyLogo" : " productSourceLogoPic")} src={this.props.item.logo} alt={this.props.item.title}/></a>
+                                </div>
+                            </div>
+                    </div>
+                </div>
+            )
+        }
+    }
 
-        let { comparedItems } = this.props;
+    render(){
 
         return(
 
 
                     <div className="row">
                         <div className="col s12 m6">
-                            <div className={" " + (comparedItems.some(item => item.title.includes(this.props.item.title)) ? "compare" : "")} >
-                                    <div className="card productCard" >
-                                            <div className="card-image itemImage" style={{display: "flex", justifyContent: "center"}}>
-                                                    <img src={this.props.item.image} alt={this.props.item.title} style={{height:"250px", width: "60%", marginTop: "25px"}}/>
-                                                    <div className="btnTool">
-                                                        <span className={!this.props.item.watch? "btnTooltipText" : ""}>{!this.props.item.watch? "Add to Watchlist" : ""}</span>
-                                                        <a className="btn-floating halfway-fab floatingWatchBtn indigo darken-4"><i className="material-icons" style={{color: (this.props.item.watch? "red" : "white")}} onClick={(this.props.item.watch) ? () => console.log("If you would like to remove this item from your watchlist, please remove it through the watchlist tab") : () => this.handleWatch(this.props.watch, this.props.item)}>{this.props.item.watch? "favorite" : "remove_red_eye"} </i></a>
-                                                    </div>             
-                                                <div className={(comparedItems.some(item => item.title.includes(this.props.item.title)) ? "card-overlay2" : "card-overlay")}></div>
-                                                <div className="detailsBtn" onClick={() => this.handleCompare(this.props.item)} style={{color: "black", display: "flex", justifyContent: "center", alignItems:"center"}}>{comparedItems.some(item => item.title.includes(this.props.item.title)) ? "Remove Compare" : "Compare"}</div>
-                                            </div>
-                                            {/* this.props.item.featured && */}
-                                            {/* <div className="featuredProduct">
-                                                Featured
-                                            </div> */}
-                                            <div className="card-content" style={{display: "flex", alignItems: "center", flexDirection: "column"}}>
-                                                <div className="card-title" style={{ lineHeight:"1.3em" ,fontSize:"18px", overflow:"hidden", position: "relative", bottom: "30%", textAlign: "center"}}>{this.props.item.title}</div>
-                                                <div className="itemPrice">{`$${this.props.item.price}`}</div>
-                                                <div>
-                                                    <a href={`${this.props.item.link}`}  target="_blank" className="productSourceLogo"><img className={"img-fluid" + (this.props.item.logo == "https://firebasestorage.googleapis.com/v0/b/poofapibackend.appspot.com/o/icons%2Fbestbuy.png?alt=media&token=da783e57-802f-4a40-b609-3af1fd78098a" ? " productBestBuyLogo" : " productSourceLogoPic")} src={this.props.item.logo} alt={this.props.item.title}/></a>
-                                                </div>
-                                            </div>
-                                    </div>
-                            </div>
+                            {this.renderCard()}
                         </div>
                     </div>
         )}
